saver/requirements.txt
paho-mqtt
psycopg[binary]~=3.2.5
